+++ androidApp/core/src/test/java/io/github/nadzwyczajnaGrupaRobocza/texaspoker/game/cards/HandComparatorAssertions.kt
@@ -1,0 +1,28 @@
+package io.github.nadzwyczajnaGrupaRobocza.texaspoker.game.cards
+
+import com.natpryce.hamkrest.assertion.assertThat
+import com.natpryce.hamkrest.equalTo
+import com.natpryce.hamkrest.greaterThan
+import com.natpryce.hamkrest.lessThan
+import compareTo
+
+fun assertHandsEqual(
+    lhs: Hand,
+    rhs: Hand
+) {
+    assertThat(lhs.compareTo(rhs), equalTo(0))
+}
+
+fun assertHandBigger(
+    lhs: Hand,
+    rhs: Hand
+) {
+    assertThat(lhs.compareTo(rhs), greaterThan(0))
+}
+
+fun assertHandLess(
+    lhs: Hand,
+    rhs: Hand
+) {
+    assertThat(lhs.compareTo(rhs), lessThan(0))
+}
